requirements.txt
jieba
numpy
pandas
Pillow
requests
xlrd
pyzbar
qrcode